+++ integration-tests/cli/test/autoinstall.test.ts
@@ -79,11 +79,22 @@ test.serial(
   }
 );
 
+/***
+ * Important: these tests are skipped because the old @openfn/language-testing package,
+ * which has been deprecated and removed from adaptors repo, depends on a version of
+ * language-common which just happens to be incompatible with node24
+ *
+ * Since the pre-release stuff is sitting unreleased these tests will just sit on skipped
+ * We should work out how to build a new version of this package (not in the adaptors repo)
+ * And probably remove the common dependency
+ */
+
 // Ignore the @next version if present but we asked for latest
-test.serial(
+test.serial.skip(
   `openfn ${jobsPath}/simple.js -a testing@latest ${repoDir} ${log}`,
   async (t) => {
     const { stdout, err } = await run(t.title);
+    console.log(stdout);
 
     // t.falsy(err); // TODO I think this is a broken adaptor build?
     t.regex(stdout, /Auto-installing language adaptors/);
@@ -109,7 +120,7 @@ test.serial(
 );
 
 // Ignore @next if present but we asked for no version
-test.serial(
+test.serial.skip(
   `openfn ${jobsPath}/simple.js -a testing ${repoDir} ${log}`,
   async (t) => {
     const { stdout, err } = await run(t.title);
@@ -126,7 +137,7 @@ test.serial(
 
 // TODO we need to fix the version of testing
 // maybe after release we can push next onto 2.0 and leave latest on 1.0
-test.serial(
+test.serial.skip(
   `openfn ${jobsPath}/simple.js -a testing@next ${repoDir} ${log}`,
   async (t) => {
     const { stdout, stderr } = await run(t.title);