+++ integration-tests/worker/src/init.ts
@@ -5,19 +5,23 @@ import createLightningServer, { toBase64 } from '@openfn/lightning-mock';
 import createEngine from '@openfn/engine-multi';
 import createWorkerServer from '@openfn/ws-worker';
 import { createMockLogger } from '@openfn/logger';
-// import createLogger from '@openfn/logger';
+import createLogger from '@openfn/logger';
+
+const debugWorker = process.env.OPENFN_DEBUG_WORKER;
+const debugLightning = process.env.OPENFN_DEBUG_LIGHTNING;
 
 export const randomPort = () => Math.round(2000 + Math.random() * 1000);
 
 export const initLightning = (port = 4000, privateKey?: string) => {
   // TODO the lightning mock right now doesn't use the secret
   // but we may want to add tests against this
-  const opts = { port };
+  const opts: any = { port };
   if (privateKey) {
-    // @ts-ignore
     opts.runPrivateKey = toBase64(privateKey);
   }
-  // opts.logger = createLogger('LTG', { level: 'debug' });
+  if (debugLightning) {
+    opts.logger = createLogger('LTG', { level: 'debug' });
+  }
   return createLightningServer(opts);
 };
 
@@ -40,8 +44,9 @@ export const initWorker = async (
   });
 
   const worker = createWorkerServer(engine, {
-    logger: createMockLogger(),
-    // logger: createLogger('worker', { level: 'debug' }),
+    logger: debugWorker
+      ? createLogger('worker', { level: 'debug' })
+      : createMockLogger(),
     port: workerPort,
     lightning: `ws://localhost:${lightningPort}/worker`,
     secret: crypto.randomUUID(),