+++ packages/runtime/src/modules/repo.ts
@@ -2,6 +2,8 @@ import path from 'node:path';
 import { readFile, writeFile, mkdir } from 'node:fs/promises';
 import { defaultLogger, Logger } from '@openfn/logger';
 import exec from '../util/exec';
+import * as os from 'node:os';
+const homeDir = os.homedir();
 
 const defaultPkg = {
   name: 'openfn-repo',
@@ -12,14 +14,13 @@ const defaultPkg = {
   dependencies: {},
 };
 
+export const defaultRepoPath = path.join(homeDir, './openfn/repo/cli');
 const npmInstallFlags = [
   '--no-audit',
   '--no-fund',
   '--no-package-lock',
   `--min-release-age=1`,
 ];
-
-export const defaultRepoPath = '/tmp/openfn/repo';
 
 type InstallList = Array<{ name: string; version: string }>;
 
@@ -241,27 +242,25 @@ export const getModuleEntryPoint = async (
       'utf8'
     );
     const pkg = JSON.parse(pkgRaw);
-    let main = 'index.js';
 
-    // TODO Turns out that importing the ESM format actually blows up
-    // (at least when we try to import lodash)
-    // if (pkg.exports) {
-    //   if (typeof pkg.exports === 'string') {
-    //     main = pkg.exports;
-    //   } else {
-    //     const defaultExport = pkg.exports['.']; // TODO what if this doesn't exist...
-    //     if (typeof defaultExport == 'string') {
-    //       main = defaultExport;
-    //     } else {
-    //       main = defaultExport.import;
-    //     }
-    //   }
-    // } else
-    // Safer for now to just use the CJS import
-    if (pkg.main) {
-      main = pkg.main;
+    // Find the best ESM entrypoint
+    // https://nodejs.org/api/packages.html#package-entry-points
+    let esm;
+    if (typeof pkg.exports === 'string') {
+      esm = pkg.exports;
+    } else {
+      const exportsField = pkg.exports?.['.'];
+      esm =
+        typeof exportsField === 'string' ? exportsField : exportsField?.import;
     }
+
+    // main might point to esm or cjs, but in our adaptors it points to CJS
+    const cjsProbably = pkg.main;
+
+    const main = esm ?? cjsProbably ?? 'index.js';
+
     const p = path.resolve(moduleRoot, main);
+
     return { path: p, version: pkg.version };
   }
   return null;